# go-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "task.shmyaks.ru API",
        "title": "InnoHack Service API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/v1/meetings/new_meeting": {
            "post": {
                "description": "Добавляет новый созвон с указанием даты и ссылки на Zoom",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meetings"
                ],
                "summary": "Создать новый созвон",
                "parameters": [
                    {
                        "description": "Запрос на создание созвона",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.CreateMeetingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Созвон успешно создан",
                        "schema": {
                            "$ref": "#/definitions/model.CodeResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка при создании созвона",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/meetings/{meeting_id}": {
            "get": {
                "description": "Возвращает детальную информацию о созвоне, включая участников",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meetings"
                ],
                "summary": "Получить данные о созвоне",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID созвона",
                        "name": "meeting_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Детали созвона",
                        "schema": {
                            "$ref": "#/definitions/model.MeetingDetailsWithParticipants"
                        }
                    },
                    "400": {
                        "description": "Ошибка при получении данных",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/meetings/{meeting_id}/invite": {
            "post": {
                "description": "Добавляет пользователя в список участников созвона",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meetings"
                ],
                "summary": "Пригласить пользователя на созвон",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID созвона",
                        "name": "meeting_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Запрос на приглашение пользователя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.InviteUserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно приглашён",
                        "schema": {
                            "$ref": "#/definitions/model.CodeResponse"
                        }
                    },
                    "400": {
                        "description": "Ошибка при приглашении пользователя",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/users/meetings": {
            "get": {
                "description": "Возвращает все созвоны, в которых пользователь является участником или создателем",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Meetings"
                ],
                "summary": "Получить все созвоны пользователя",
                "responses": {
                    "200": {
                        "description": "Список созвонов",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.MeetingDetails"
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при получении данных",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.CodeResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "model.CreateMeetingRequest": {
            "type": "object",
            "properties": {
                "meeting_date": {
                    "type": "string"
                },
                "zoom_link": {
                    "type": "string"
                }
            }
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "model.InviteUserRequest": {
            "type": "object",
            "properties": {
                "user_name": {
                    "type": "string"
                }
            }
        },
...
        "model.MeetingDetails": {
            "type": "object",
            "properties": {
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "meeting_date": {
                    "type": "string"
                },
                "zoom_link": {
                    "type": "string"
                }
            }
        },
        "model.MeetingDetailsWithParticipants": {
            "type": "object",
            "properties": {
                "created_by": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "meeting_date": {
                    "type": "string"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "zoom_link": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "CookieAuth": {
            "type": "apiKey",
            "name": "Authtoken",
            "in": "cookie"
        }
    }
}